+++ src/main/java/com/project/trysketch/dto/request/GameRoomRequestDto.java
@@ -1,0 +1,12 @@
+package com.project.trysketch.dto.request;
+
+import lombok.AllArgsConstructor;
+import lombok.Getter;
+import lombok.NoArgsConstructor;
+
+@Getter
+@AllArgsConstructor
+@NoArgsConstructor
+public class GameRoomRequestDto {
+    private String title;
+}
