+++ src/main/java/com/project/trysketch/dto/response/GameRoomResponseDto.java
@@ -1,0 +1,35 @@
+package com.project.trysketch.dto.response;
+
+import com.fasterxml.jackson.annotation.JsonFormat;
+import lombok.AllArgsConstructor;
+import lombok.Builder;
+import lombok.Getter;
+import lombok.NoArgsConstructor;
+import org.springframework.data.annotation.CreatedDate;
+import org.springframework.data.annotation.LastModifiedDate;
+
+import javax.persistence.Column;
+import java.time.LocalDateTime;
+import java.util.List;
+
+// 1. 기능   : 게임 방 response용 dto
+// 2. 작성자 : 김재영
+@Getter
+@NoArgsConstructor
+@Builder
+@AllArgsConstructor
+public class GameRoomResponseDto {
+    private Long id;
+    private String title;
+    private int GameRoomUserCount;
+    private String host;
+    private String status;
+
+    @JsonFormat(shape = JsonFormat.Shape.STRING, pattern = "yyyy-MM-dd HH:mm:ss", timezone = "Asia/Seoul")
+    private LocalDateTime createdAt;
+
+    @JsonFormat(shape = JsonFormat.Shape.STRING, pattern = "yyyy-MM-dd HH:mm:ss", timezone = "Asia/Seoul")
+    private LocalDateTime modifiedAt;
+
+
+}
