+++ src/main/java/com/project/trysketch/dto/response/UserResponseDto.java
@@ -1,0 +1,10 @@
+package com.project.trysketch.dto.response;
+
+import lombok.Getter;
+
+@Getter
+public class UserResponseDto {
+    private Long id;
+    private String email;
+    private String nickname;
+}
