--- src/main/java/com/project/trysketch/gameroom/dto/request/GameRoomRequestDto.java
@@ -1,12 +1,0 @@
-package com.project.trysketch.gameroom.dto.request;
-
-import lombok.AllArgsConstructor;
-import lombok.Getter;
-import lombok.NoArgsConstructor;
-
-@Getter
-@AllArgsConstructor
-@NoArgsConstructor
-public class GameRoomRequestDto {
-    private String title;
-}
