+++ src/main/java/com/project/trysketch/image/ImageService.java
@@ -4,9 +4,8 @@ import com.project.trysketch.global.dto.MsgResponseDto;
 import com.project.trysketch.global.exception.CustomException;
 import com.project.trysketch.global.exception.StatusMsgCode;
 import com.project.trysketch.global.jwt.JwtUtil;
-import com.project.trysketch.global.rtc.Room;
-import com.project.trysketch.user.entity.User;
-import com.project.trysketch.user.repository.UserRepository;
+import com.project.trysketch.entity.User;
+import com.project.trysketch.repository.UserRepository;
 import io.jsonwebtoken.Claims;
 import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Service;
