+++ src/main/java/com/project/trysketch/redis/service/RedisService.java
@@ -4,10 +4,9 @@ import com.project.trysketch.global.jwt.JwtUtil;
 import com.project.trysketch.redis.entity.Guest;
 import com.project.trysketch.redis.dto.GuestNickRequestDto;
 import com.project.trysketch.redis.repositorty.GuestRepository;
-import com.project.trysketch.user.service.UserService;
+import com.project.trysketch.service.UserService;
 import lombok.RequiredArgsConstructor;
 import org.springframework.beans.factory.annotation.Value;
-import org.springframework.cache.annotation.Cacheable;
 import org.springframework.data.redis.core.RedisTemplate;
 import org.springframework.stereotype.Service;
 import javax.servlet.http.HttpServletRequest;
