+++ src/main/java/com/project/trysketch/service/KakaoService.java
@@ -1,0 +1,151 @@
+package com.project.trysketch.service;
+
+import com.fasterxml.jackson.core.JsonProcessingException;
+import com.fasterxml.jackson.databind.JsonNode;
+import com.fasterxml.jackson.databind.ObjectMapper;
+import com.project.trysketch.global.jwt.JwtUtil;
+import com.project.trysketch.dto.request.KakaoUserRequstDto;
+import com.project.trysketch.entity.User;
+import com.project.trysketch.repository.UserRepository;
+import org.springframework.http.HttpEntity;
+import org.springframework.http.HttpHeaders;
+import org.springframework.http.HttpMethod;
+import org.springframework.http.ResponseEntity;
+import org.springframework.security.crypto.password.PasswordEncoder;
+import org.springframework.stereotype.Service;
+import org.springframework.util.LinkedMultiValueMap;
+import org.springframework.util.MultiValueMap;
+import org.springframework.web.client.RestTemplate;
+import javax.servlet.http.HttpServletResponse;
+import lombok.RequiredArgsConstructor;
+import lombok.extern.slf4j.Slf4j;
+import java.util.UUID;
+
+// 1. 기능    : OAuth2.0 카카오 비즈니스 로직
+// 2. 작성자  : 황미경
+
+@Slf4j
+@Service
+@RequiredArgsConstructor
+public class KakaoService {
+
+    private final PasswordEncoder passwordEncoder;
+    private final UserRepository userRepository;
+    private final JwtUtil jwtUtil;
+    private final UserService userService;
+
+    public void kakaoLogin(String code, HttpServletResponse response) throws JsonProcessingException {
+        String randomNickname = userService.RandomNick();
+
+        // 1. "인가 코드"로 "액세스 토큰" 요청
+//        String accessToken = getToken(code);                                           // 포스트맨 확인위해 주석처리 필요
+
+        // 2. 토큰으로 카카오 API 호출 : "액세스 토큰"으로 "카카오 사용자 정보" 가져오기
+        KakaoUserRequstDto kakaoUserInfo = getKakaoUserInfo(code, randomNickname);             // 포스트맨 확인위해 accessToken에서 code로 바꿔야함
+
+        // 3. 필요시에 회원가입
+        User kakaoUser = registerKakaoUserIfNeeded(kakaoUserInfo);
+
+        // 4. JWT 토큰 반환
+        String createToken =  jwtUtil.createToken(kakaoUser.getEmail(), randomNickname);
+        response.addHeader(JwtUtil.AUTHORIZATION_HEADER, createToken);
+    }
+
+    // 1. "인가 코드"로 "액세스 토큰" 요청
+    private String getToken(String code) throws JsonProcessingException {
+        // HTTP Header 생성
+        HttpHeaders headers = new HttpHeaders();
+        headers.add("Content-type", "application/x-www-form-urlencoded;charset=utf-8");
+
+        // HTTP Body 생성
+        MultiValueMap<String, String> body = new LinkedMultiValueMap<>();
+        body.add("grant_type", "authorization_code");
+        body.add("client_id", "99896cbca8689b2a7b2513df031382da");
+        
+        // body.add("redirect_uri", "http://localhost:8080/api/user/kakao/callback");  // 포스트맨 실험
+        
+        body.add("redirect_uri", "http://localhost:3030/login");                       // 프론트의 주소
+        body.add("code", code);
+
+        // HTTP 요청 보내기
+        HttpEntity<MultiValueMap<String, String>> kakaoTokenRequest = new HttpEntity<>(body, headers);
+        RestTemplate rt = new RestTemplate();
+        ResponseEntity<String> response = rt.exchange(
+                "https://kauth.kakao.com/oauth/token",
+                HttpMethod.POST,
+                kakaoTokenRequest,
+                String.class
+        );
+
+        // HTTP 응답 (JSON) -> 액세스 토큰 파싱
+        String responseBody = response.getBody();
+        ObjectMapper objectMapper = new ObjectMapper();
+        JsonNode jsonNode = objectMapper.readTree(responseBody);
+        return jsonNode.get("access_token").asText();
+    }
+
+    // 2. 토큰으로 카카오 API 호출 : "액세스 토큰"으로 "카카오 사용자 정보" 가져오기
+    private KakaoUserRequstDto getKakaoUserInfo(String accessToken, String randomNickname) throws JsonProcessingException {
+        // HTTP Header 생성
+        HttpHeaders headers = new HttpHeaders();
+        headers.add("Authorization", "Bearer " + accessToken);
+        headers.add("Content-type", "application/x-www-form-urlencoded;charset=utf-8");
+
+        // HTTP 요청 보내기
+        HttpEntity<MultiValueMap<String, String>> kakaoUserInfoRequest = new HttpEntity<>(headers);
+        RestTemplate rt = new RestTemplate();
+        ResponseEntity<String> response = rt.exchange(
+                "https://kapi.kakao.com/v2/user/me",
+                HttpMethod.POST,
+                kakaoUserInfoRequest,
+                String.class
+        );
+
+        String responseBody = response.getBody();
+        ObjectMapper objectMapper = new ObjectMapper();
+        JsonNode jsonNode = objectMapper.readTree(responseBody);
+        Long id = jsonNode.get("id").asLong();
+        String nickname = jsonNode.get("properties")
+                .get("nickname").asText();
+        String email = jsonNode.get("kakao_account")
+                .get("email").asText();
+
+        log.info("카카오 사용자 정보: " + id + ", " + nickname + ", " + email);
+        log.info("저장할 사용자 정보 : " + id + ", " + randomNickname + ", " + email);
+        return new KakaoUserRequstDto(id, randomNickname, email);
+    }
+
+    // 3. 필요시에 회원가입
+    private User registerKakaoUserIfNeeded(KakaoUserRequstDto kakaoUserInfo) {
+        // DB 에 중복된 Kakao Id 가 있는지 확인
+        Long kakaoId = kakaoUserInfo.getId();
+        User kakaoUser = userRepository.findByKakaoId(kakaoId)
+                .orElse(null);
+        if (kakaoUser == null) {
+            // 카카오 사용자 email 동일한 email 가진 회원이 있는지 확인
+            String kakaoEmail = kakaoUserInfo.getEmail();
+
+            User sameEmailUser = userRepository.findByEmail(kakaoEmail).orElse(null);
+
+            if (sameEmailUser != null) {
+                kakaoUser = sameEmailUser;
+                // 기존 회원정보에 카카오 Id 추가
+                kakaoUser = kakaoUser.kakaoIdUpdate(kakaoId);
+            } else {
+                // 신규 회원가입
+                String password = UUID.randomUUID().toString();
+                String encodedPassword = passwordEncoder.encode(password);
+                String email = kakaoUserInfo.getEmail();
+
+                kakaoUser = User.builder()
+                        .password(encodedPassword)
+                        .kakaoId(kakaoId)
+                        .kakaoNickname(kakaoUserInfo.getNickname())
+                        .email(email)
+                        .build();
+            }
+            userRepository.save(kakaoUser);
+        }
+        return kakaoUser;
+    }
+}
