--- src/main/java/com/project/trysketch/user/dto/UserResponseDto.java
@@ -1,10 +1,0 @@
-package com.project.trysketch.user.dto;
-
-import lombok.Getter;
-
-@Getter
-public class UserResponseDto {
-    private Long id;
-    private String email;
-    private String nickname;
-}
